create PermissionTemplate for imported Collections with manager
When creating a Collection through the UI, the creator is automatically
assigned to the new Collection as a manager through the Collection's
PermissionTemplate. Currently, it seems like imported Collections don't
get a PermissionTemplate. We should create a PermissionTemplate for the
Collection if one doesn't already exist and add the User who created the
Importer as a manager of that Collection.