Imported Collections get a PermissionTemplate with a manager (#245)
* create PermissionTemplate for imported Collections with manager

When creating a Collection through the UI, the creator is automatically
assigned to the new Collection as a manager through the Collection's
PermissionTemplate. Currently, it seems like imported Collections don't
get a PermissionTemplate. We should create a PermissionTemplate for the
Collection if one doesn't already exist and add the User who created the
Importer as a manager of that Collection.

* call #reset_access_controls! after creating PermissionTemplateAccess

* always create a PermissionTemplateAccess

this replaces using manage_users on PermissionTemplate creation, which is invalid

Co-authored-by: Daniel Pierce <dlpierce@indiana.edu>

* fix trailing whitespace

Co-authored-by: Daniel Pierce <dlpierce@indiana.edu>