Update metadata only handling (#336)
* exporting numerated headers

* comments and improvements

* rubocop

* fix lint issue, revert accidental version bump

* stops file set id from being prepended to exported file names if metadata only

* remove unneccessary conditional

* add spec for the export_headers method

* rubocop fixes

* rubocop again

* more file skips if metadata only

* only loop through the entries once
co-authored by: rob kaufman: <rob@notch8.com>

* save entry headers to self

* csv parser specs pass again

* remove access_control_id as a header

* better filtering

* better csv

* Refactor csv export to use from values instead of to values.

CSV export was using the wrong headers. If you specify `"title" => {from: ['display_title']}` in the mapper,
then you would expect the export to say `display_title` and not `title`. This was not the case.

This commit also adds a new option for the field mapping - join. The default behavior on export has been to
join fields together, but field_1,field_2 easier for users to edit. This will be the new default and join allows
you to specify the old behavior.

Also cleaned up object code on exporter and reduced cyclomatic complexity on some methods in csv entry.

* include object headers again (#340)

* include object headers again

* rubocop

* use "with_indifferent_access" and fix specs

* remove comment

* Update csv_parser.rb

* Update csv_parser.rb

* update value indexing

* update fedora query

* rubocop

* title_1 or title_2 are just as good as title

* fix import validator

* make sure object fields have index

* allow id if matched

* get rid of "undefined method id for array..." error

* return the ids from the query

* refines metadata_only checking

* removed old `if else` formatting since it doesn't make sense anymore

* update csv parser

* lint fix

* lint fix 2

Co-authored-by: Alisha Evans <alisha@notch8.com>
Co-authored-by: Rob Kaufman <rob@notch8.com>
Co-authored-by: Alisha Evans <alishaevn2@gmail.com>