Ensuring metadata_prefix option is correctly passed (#677)
Prior to this commit, the `metadata_prefix` passed on `OAI::Client.new`
was discarded.  In the current code, that option is not processed.  It
is instead a parameter to pass to the OAI verb
action (e.g. `ListIdentifiers` and `GetRecords`).

With this commit, we ensure that we pass the `metadata_prefix` to the
specific verb.

See
https://github.com/code4lib/ruby-oai/blob/d0303be9800e7247532286330cf3b0cf135f19d7/lib/oai/client.rb#L86-L121

Closes https://github.com/samvera-labs/bulkrax/issues/676

list#