Normalizing serialized data for BOM characters
Prior to this commit, when someone would upload a CSV it might have a
column name that included the tricksy, invisibile [Byte Order Mark][1].
The manifestation was that you could look at the raw metadata and see a
key called `"file"` however when checking `raw_metadata.key?("file")`
the result was `false`.

To find this `entry.raw_metadata.keys.first.chars.map(&:chr)` I got
`["", "f", "i", "l", "e"]` where the first value of that array was a
[Byte Oder Mark][1].

With this commit, we're handling both how we persist and how we load
persisted serialized data.  This follows on the documentation for the
`ActiveRecord::Base.serialize` method.

It's envisioned that we might have more characters to sanitize.

Closes: samvera-labs/bulkrax#688
Related to: scientist-softserv/adventist-dl#179

[1]: https://en.wikipedia.org/wiki/Byte_order_mark