Ensuring OAI parser sets factory class then parses (#703)
* Ensuring OAI parser sets factory class then parses

Let's say we have nodes with the following names (and order): `shape`, `model`,
`smell`

Prior to this commit, as we were reading through the record's metadata
when we were processing `shape` we would use the default `factory_class`
to check if `shape` was a valid field.  Then we'd use the value of
`model` to derive the `factory_class`.  And when we then processed
`smell` we'd be using the newly derived `factory_class` from the `model`
node.

With this commit, we first establish the `factory_class` from the
`model` node.  Then we again process the all the nodes to parse the
metadata.  This ensures that both `shape` and `smell` are parsed
against the expected `factory_class`.

Further, I have made a few refactorings that should ease upstream
overrides by extracting methods from the body of the `build_metadata`
method.

Related to:

- https://github.com/samvera-labs/bulkrax/issues/702
- https://github.com/scientist-softserv/adventist-dl/issues/188
- https://github.com/scientist-softserv/adventist-dl/issues/187

* Removing comment that no longer applies

* Removing `break` in loop for setting factory class

In discussions with Kiah, having the `break` adds a slight performance
improvement.  However the `break` did break with the existing pattern.

We do take efforts to set the factory_class to the first encountered
`model`, so having mutliple `models` is probably okay.