Ensuring XML parser sets factory class then parses
This commit for changing the `Bulkrax::XmlEntry` is the sibling to the
changes made to `Bulkrax::OaiEntry` (see SHA 35970cee).  The salient
information is pulled as a quote from that commit:

> Prior to this commit, as we were reading through the record's metadata
> when we were processing `shape` we would use the default
> `factory_class` to check if `shape` was a valid field.  Then we'd use
> the value of `model` to derive the `factory_class`.  And when we then
> processed `smell` we'd be using the newly derived `factory_class` from
> the `model` node.
>
> With this commit, we first establish the `factory_class` from the
> `model` node.  Then we again process the all the nodes to parse the
> metadata.  This ensures that both `shape` and `smell` are parsed
> against the expected `factory_class`.

In addition, I have refactored the spec to remove global state pollution
of the parser config.

I have also renamed and documented a method to be more descriptive; the
previous method name of `xml_elements` was opaque, as it returned an
Array of strings (and not an array of XML elements).

Further steps could be to extract the now "ubiquitous" `Avacado` class
into a spec fixture.

We may want to revisit the XML parser fundamental logic; namely we only
parse nodes that are explicitly declared with in the `from`.  This is a
bit different than other parsers, in that they will make assumptions
about each encountered column (in the case of CSV) or node (in the case
of OAI).  tl;dr - Here there be dragons.

Closes #702

Related to:

- https://github.com/samvera-labs/bulkrax/pull/703
- https://github.com/scientist-softserv/adventist-dl/issues/188
- https://github.com/scientist-softserv/adventist-dl/issues/187
- https://github.com/scientist-softserv/adventist-dl/issues/186

[1]: