Scour blank strings from the creation attributes
There are three changes in this commit:

1. Exposing a class attribute
2. Scouring attribute hash
3. Extracting a method

**Exposing a class attribute**

This was added to allow for easier testing of a private method.  It also
"exposes" a hidden assumption in a private method; turning that
assumption into a telegraphed potential configuration point.

**Scouring attribute hash**

This resolves the reported problem by @eporter23:

> When testing an import on 9/16/22, the work row failed to validate,
> returning an error regarding "invalid entry for administrative_unit".
>
> The CSV used had the header for that field, but no values entered. Our
> preprocessors output field headers for mapped fields, but sometimes
> there are no values entered for a given header.
>
> -- https://github.com/emory-libraries/dlp-curate/issues/1973

The likely most vexing problem being an array that has only empty
values.  This is likely something that should be remedied upstream in
the data modeling for Hyrax models; but alas we're not there yet.

**Extracting a method**

This is a happy little refactor to create symmetry between the other
create/update methods for collections and file sets.  In doing this, we
provide a mechanism for folks to more readily over-ride methods at a
lower level.

Related to:

- https://github.com/emory-libraries/dlp-curate/issues/1973
- https://github.com/emory-libraries/dlp-curate/pull/1981