Adding ApplicationParser#file_set_entry_class to API
Prior to this commit, this method was not defined on the
`ApplicationParser` yet there was an assumption in the
`Bulkrax::ImportersController` and `Bulkrax::ExportersController` that
this method existed.

With this commit, we're acknowledging that each parser should implement
this method.

In addition this commit adds a `#create_file_sets` method for the
`Bulkrax::XmlParser`.  This change mirrors the
`Bulkrax::XmlParser#create_collections` implementation and todo item.

A future state that would be wonderful is to create spec helpers that
implementers could use in their specs to ensure a compliant interface.
This echoes how Valkyrie does things.