Fixing bug introduced in #663
Prior to this commit, I had refactored (in PR #663) logic to remove
conditional logic.  However, I had introduced a bug, namely using
`parser` and the receiver of `send`.  For the
`Bulkrax::ApplicationParser` there is no parser method.  This resulted
in a notable bug.

With this commit, I'm removing `parser` as the receiver of `send` and
instead favoring the implicit `self` as the receiver of `send`.