Ensuring CSV sets factory class before metadata parsing (#719)
* Ensuring CSV sets factory class before metadata parsing

Prior to this commit, if we had configured bulkrax such that we were
allowing for our model import to come from multiple fields
(e.g. `from: ['model', 'work_type']`) and the CSV had `work_type` as the
column name, we would get an error as follows:

```
NoMethodError:
  undefined method `method_defined?' for nil:NilClass
    # ./app/models/concerns/bulkrax/has_matchers.rb:128:in `field_supported?'
    # ./app/models/concerns/bulkrax/has_matchers.rb:157:in `multiple?'
    # ./app/models/concerns/bulkrax/has_matchers.rb:34:in `block in add_metadata'
    # ./app/models/concerns/bulkrax/has_matchers.rb:31:in `each'
    # ./app/models/concerns/bulkrax/has_matchers.rb:31:in `add_metadata'
    # ./app/models/bulkrax/csv_entry.rb:77:in `block in add_ingested_metadata'
    # ./app/models/bulkrax/csv_entry.rb:75:in `each'
    # ./app/models/bulkrax/csv_entry.rb:75:in `add_ingested_metadata'
    # ./app/models/bulkrax/csv_entry.rb:43:in `build_metadata'
```

Tracing this line it is because the `factory_class` is nil.  And if the
model column were to be a later column, any columns that occurred before
we established the factory class may have been skipped.

With this commit, we're echoing the functionality introduced in #705 and #703.

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/211
- https://github.com/samvera-labs/bulkrax/pull/705
- https://github.com/samvera-labs/bulkrax/pull/703

Closes #718

* Update spec/fixtures/csv/factory_class_test.csv

Co-authored-by: Benjamin Kiah Stroud <32469930+bkiahstroud@users.noreply.github.com>

Co-authored-by: Benjamin Kiah Stroud <32469930+bkiahstroud@users.noreply.github.com>