Adding config to reindex child work's file sets
Prior to this commit, we did not expose a means to cascade indexing
deeper than the child work.

With this commit we provide a configuration point for saying "Please
reindex the file sets of the child works we're processing for this
relationship job.

Closes #721

Related to:

- https://github.com/scientist-softserv/louisville-hyku/commit/128a9ef

**Release Notes**

If you wish to enable reindex child works's file sets during the
CreateRelationshipsJob, you'll want to add the following to your
`config/initializer/bulkrax.rb`:

```ruby
Bulkrax::CreateRelationshipsJob.update_child_records_works_file_sets = true
```