Extract memory performant export record record set
Prior to this commit the `Bulkrax::CsvParser` had complex query logic
for finding the records to process for export.  The logic was memory
inefficient: calling `#current_record_ids` twice (a very expensive
queries that issued many calls to Fedora which could've been a SOLR
query).  In large exports, the processes would spend over 90% of the
time in the finding file sets.

With this commit the `Bulkrax::ParserExportRecordSet` finder logic
echoes the logic once found in `Bulkrax::CsvParser`.  The purpose is to
isolate the place where the SOLR queries happen and to create less
memory intensive processes.

Further the `Bulkrax::ParserExportRecordSet.for` returns an object that
"quacks" like an Array of Arrays, with each inner Array having two
elements.  This makes the testing a bit easier.

Further, this extraction helps us create the conditions for other
exporters to re-use the export from logic that was presently tucked away
in the CSV Parser.

Of particular note, there are no real SOLR query tests of these changes,
nor the code prior to this commit.

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/289