Adding logic to skip blank lines
As we've been working with folks uploading CSVs, we're encountering
issues when they upload CSVs that has lines that don't have data.

There are three types of lines we might encounter:

- a blank line; the singular line looks as follows `\n`
- a line with empty fields; the singular line looks as follows `,,,\n`
- a line might might have empty strings for cells (e.g. `"","","",""\n`)

These blank lines are making their way to the parser's `@records`
instance variable and then we start seeing weird exceptions such as the
following:

```
Missing identifier for {:﻿file=>nil, :identifier=>nil,
:"identifier.ark"=>nil, :title=>nil, :description=>nil, :creator=>nil,
:contributor=>nil, :date=>nil, :"date.other"=>nil,
:"format.extent"=>nil, :type=>nil, :subject=>nil, :language=>nil,
:source=>nil, :"relation.isPartOf"=>nil, :rights=>nil,
:"coverage.spatial"=>nil}
```

There are no values in the above hash, so my working assumption is that
the row is one of the aforementioned types of rows that are causing
havoc.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/113
- https://github.com/scientist-softserv/britishlibrary/issues/280