Avoid building value/object if hyrax_record does not respond
Prior to this commit, we'd attempt to export the field even if the
"from" object did not respond to that property.  That would raise a
`NoMethodError`.

With this commit, we're skipping the building of the value/object if the
hyrax_record does not respond to the source method.

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/289
- https://github.com/samvera-labs/bulkrax/issues/754