Fixing bug regarding total and export limits
The `current_record_ids` was removed in 4cf0207 (e.g. #749) in favor of
`current_records_for_export`; however the `Bulkrax::CsvParser#total`
method was not updated to reflect this change.

In this commit, we fix that issue.  Also we tighten up the logic
regarding the sorted_entries array.  There was an assumption that we
would have a `nil` limit; but the documentation mentions that 0 and
`nil` are equivalent.  But `(0..(0 || total))` is different from
`(0..(nil || total))`.

Partially closes #754