I754 (#756)
* Avoid building value/object if hyrax_record does not respond

Prior to this commit, we'd attempt to export the field even if the
"from" object did not respond to that property.  That would raise a
`NoMethodError`.

With this commit, we're skipping the building of the value/object if the
hyrax_record does not respond to the source method.

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/289
- https://github.com/samvera-labs/bulkrax/issues/754

* Fixing bug regarding total and export limits

The `current_record_ids` was removed in 4cf0207 (e.g. #749) in favor of
`current_records_for_export`; however the `Bulkrax::CsvParser#total`
method was not updated to reflect this change.

In this commit, we fix that issue.  Also we tighten up the logic
regarding the sorted_entries array.  There was an assumption that we
would have a `nil` limit; but the documentation mentions that 0 and
`nil` are equivalent.  But `(0..(0 || total))` is different from
`(0..(nil || total))`.

Partially closes #754