add rights_statement to list of fields with multiple values
By default, `ImportBehavior#add_rights_statement` sets the value to an array,
which causes an error since `sanitize_controlled_uri_values!` doesn't know to
treat it as such.

This whole section needs to be refactored, since the rights statement service
can be overridden locally, but this at least un-breaks the default configuration.