Adding `Samvera::Derivatives` for interface discussion
This commit begins to define an interface for the separation of concerns
regarding the long-standing implementation of `Hydra::Derivatives` and
`Hyrax::FileSetDerivativesService` as those relate to the
existence of pre-existing derivatives.

This does not address the implementation details related to Bulkrax
handling this work.  I would instead assume that Bulkrax would implement
a subclass(es) `Samvera::Derivatives::FileLocator::Strategy`, which
would in turn address the specific needs and implementation details.

As I've explored the implementations, a challenge of Hydra::Derivatives
is that is combines the concern of finding the derivative (by making it)
and applying that derivative to the corresponding file set.

Also, there is a higher level concern that is not addressed in this
implementation; namely what are the derivatives of concern for a
given `FileSet`; something that is likely based on the mime_type of the
original file and metadata from the parent work.

Why not create another gem?

1. The immediate need is Bulkrax related.
2. The plumbing is laid for Bulkrax's test suite.
3. We probably don't need another gem to triangulate for a general
   solution.

That said, I am name spacing this interface in `Samvera` so if we find
it useful we can extract a `samvera-derivatives` gem.  After all, not
everyone's going to want Bulkrax.

Why not put this in Hyrax?

1. Because Bulkrax works across many different Hyrax versions.  Thus by
   being in Bulkrax it is generally more available than if it were in
   Hyrax; something which we'd need to backport.

An advantage of establishing the interface is that work can be done to
now "fan out".  That is someone can work on a locator strategy while
another works on the applicator strategy.

Related to:

- https://github.com/samvera-labs/bulkrax/issues/760
- https://github.com/scientist-softserv/utk-hyku/issues/343