Adding `Bulkrax::EntryDerivative` model
This is preliminary modeling for imports to handle associating existing
derivatives with a `Bulkrax::Entry`.

Hyrax has implicit derivative_types that are not declared, but in part
assumed as a general list of derivatives.  In this model, we want to
begin name those derivatives to address the following use cases:

```gherkin
Given a FileSet
When I provide a thumbnail derivative
Then I want to add that as the thumbnail for the FileSet

Given a FileSet
When I do not provide a thumbnail derivative
Then I want to generate a thumbnail
And add the generated as the thumbnail for the FileSet
```

Related to:

- https://github.com/samvera-labs/bulkrax/pull/761
- https://github.com/samvera-labs/bulkrax/issues/760
- https://github.com/scientist-softserv/utk-hyku/issues/371