Ensuring we handle non-string keys for persisted mappings (#771)
* Ensuring we handle non-string keys for persisted mappings

Prior to this commit, the `mapping` method is data persisted in the
`importerexporter`.  Those are persisted and reified with string keys.

We discovered this bug in a failing spec for
https://github.com/scientist-softserv/britishlibrary/pull/360

With this commit, we're coercing the keys to strings and then checking
the stringified keys.  This ensures that if the implementation of
`mapping` (casting the keys to symbols, which is how we write them in
the parsers) were to change we are insulated from that change.

Related to:

- https://github.com/samvera-labs/bulkrax/pull/765

* Update app/parsers/bulkrax/csv_parser.rb

---------

Co-authored-by: Shana Moore <shana.lavina.moore@gmail.com>