Addressing SOLR maxBooleanClauses for WorkType export (#776)
Prior to this commit, we were querying all of the candidate file set IDs
for the associated work type.  In the case fo the below exporter, we had
1196 IDs to query.

As we were testing if there was a query size limit, we noticed that we
could query 1024 IDs but not 1025.  We add other query parameters, to
bump the query size, but at 1024 IDs things worked.  At 1025 they
failed.

Digging further, we saw that there is hard (yet configurable) limit for
SOLR query OR clauses.  In our case, 1024.  (Hint: if you powers of 2,
dig a little).

With this commit, we instead request 512 IDs at a time and join those
results into a single array.

Note: there are other queries that will need similar fixing but for the
time being, this (we hope) resolves a client urgency.

See: https://kew.iro.bl.uk/exporters/146?locale=en
See: https://solr.apache.org/guide/8_1/query-settings-in-solrconfig.html#query-sizing-and-warming

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/289
- https://github.com/samvera-labs/bulkrax/pull/749
- https://github.com/samvera-labs/bulkrax/issues/775