HotFix: Patching Marcel MimeType sniffing (#778)
Note: This does not included tests, which is something that #777 also
did not.  However, we have verified the behavior.

Reading the [documentation for the Marcel gem][1], it requires a file
handle (e.g. `File.new` or `File.open`) instead of a String.

When given a `String` for `::Marcel::MimeType.for` the returned value is
`application/octet-stream`.

Related to:

- https://github.com/samvera-labs/bulkrax/issues/777

[1]: https://github.com/rails/marcel