move config to a Configurable class (#781)
* move config to a Configurable class

the older style config setup is incompatible with Rails 6, preventing support
for Hyrax 4.

this is a minimal pass at extracting the config and keeping the current API.

* disable rubocop Style/ModuleFunction for Bulkrax module

there's some debate about the value of this style guide rule at
https://github.com/rubocop/ruby-style-guide/issues/556.

it seems in our case like using `module_function` is going to be prohibitively
hard, and i think the longer term desire is to remove both `extend Forwardable`
and `extend self`, expecting callers to just call `#config` directly. the
behavior around defaults needs to be cleaned up to make this possible, but it
seems like a bad plan to bend this implementation around (unhelpful?) style
rules for code we'd rather work on removing.