🐛 Reindex child work after adding relationship
Changes child `save!` to `update_index` and moves it to after the
parent record is saved.

Parent/child relationships are stored on the parent, but indexed
onto the child. A child work will not have the correct relationships
saved in the index if not reindexed after the parent is saved with the
relationship.

Refs https://github.com/samvera-labs/bulkrax/issues/810