⚙️ Removing yaml aliases from test database config
In specs, I was seeing the following error:

```
Psych::AliasesNotEnabled: Cannot load database configuration: Alias
parsing was not enabled. To enable it, pass `aliases: true` to
`Psych::load` or `Psych::safe_load`.
```

Given that we rarely touch the file, I see no benefit to using an alias
relative to the above error exception.