🐛 Ensuring we inject Bulkrax views before Hyrax
Prior to this commit, the logic for finding the Hyrax in the view path
was assuming a version number (e.g. `hyrax-2.`); however if the we
pinned Hyrax to a branch/sha then we might have `hyrax-a`.  The result
being that we pre-prended the view path into the application.  That
created the conditions where the `view_paths` first element might not
have been the application but instead a gem/engine.

Which means that the convention of overriding views in the application
would not work.

With this commit, we're using the Hyrax's engine's root to determine the
view path suffix.  Further, if Hyrax is not in the view path, we don't
again inject at the beginning of the list the Bulkrax engine into the
view path.

Related to:

- https://github.com/scientist-softserv/palni-palci/pull/731