🐛 Fix tabs for Hydra application (#875)
This commit will add logic for showing the correct tab in a Hydra
application like WVU's.  We found that if we add the bootstrap require
in their application.js then Blacklight's dropdown menu's stop working.
However this then breaks Bulkrax tabs.  Because we're not adding
bootstrap into the Hydra appilication, we mimic the behavior of the tabs
function instead.