add support for Rails 6, Hyrax 4, and Blacklight 7 (#782)
* add support for Rails 6, Hyrax 4, and Blacklight 7

this makes no attempt to ensure UI-layer support for BL7 and whatever bootstrap
versions it requires (excepting whatever is tested by RSpec). the scope here is
simply to make it *possible* for a Hyrax 4 application to install and use
Bulkrax.

* include current Hyrax CSS class and onclick in sidebar additions

* change `.hidden` css classes to `.d-none`

* bootstrap upgrade nav-link classes

* update dry-monads to ~> 1.5.0

* re-run rubocop

* begin testing on ruby 3, drop support for 2.6

* no more auto foreign key

* update schema in test app

* add support for Rails 6, Hyrax 4, and Blacklight 7

this makes no attempt to ensure UI-layer support for BL7 and whatever bootstrap
versions it requires (excepting whatever is tested by RSpec). the scope here is
simply to make it *possible* for a Hyrax 4 application to install and use
Bulkrax.

* include current Hyrax CSS class and onclick in sidebar additions

* change `.hidden` css classes to `.d-none`

* bootstrap upgrade nav-link classes

* update dry-monads to ~> 1.5.0

* re-run rubocop

* begin testing on ruby 3, drop support for 2.6

* no more auto foreign key

* update schema in test app

* trying to talk ci in to passing

* add support for Rails 6, Hyrax 4, and Blacklight 7

this makes no attempt to ensure UI-layer support for BL7 and whatever bootstrap
versions it requires (excepting whatever is tested by RSpec). the scope here is
simply to make it *possible* for a Hyrax 4 application to install and use
Bulkrax.

* include current Hyrax CSS class and onclick in sidebar additions

* change `.hidden` css classes to `.d-none`

* bootstrap upgrade nav-link classes

* update dry-monads to ~> 1.5.0

* re-run rubocop

* begin testing on ruby 3, drop support for 2.6

* no more auto foreign key

* update schema in test app

* trying to talk ci in to passing

* deal with differences in rails spec db handling for rails 6

---------

Co-authored-by: Alexandra Dunn <dunn.alex@gmail.com>
Co-authored-by: Rob Kaufman <rob@notch8.com>