I897 Bulkrax readiness for Hyku 6 and Hyrax 4 & 5 (#898)
* :broom: relocates transactions from inititalizer file

Issue:
- https://github.com/samvera/bulkrax/issues/897

Co-Authored-By: LaRita Robinson <laritakr@users.noreply.github.com>

* :broom: Add specs for container.rb, relocate files

Co-Authored-By: LaRita Robinson <laritakr@users.noreply.github.com>

* :broom: normalize magic strings into constants for referencing later

Convert the create_with_bulk_behavior and update_with_bulk_behavior to a constant; that way we can reference it in IiifPrint and document the “magic” string.

Co-Authored-By: LaRita Robinson <laritakr@users.noreply.github.com>

* :broom: correct camel case to constant notation for easier referencing

Co-Authored-By: LaRita Robinson <laritakr@users.noreply.github.com>

* :lipstick: rubocop fixes

Co-Authored-By: LaRita Robinson <laritakr@users.noreply.github.com>

* Update app/factories/bulkrax/valkyrie_object_factory.rb

* Update spec/bulkrax/transactions/container_spec.rb

* 🧹 Move container & steps

Match Hyrax convention by using bulkrax/transactions.

* restructure org to run specs locally

receiving error when trying to run the entire spec suite due to restructuring files but not moving the spec file.

* 🚧 WIP: Consolidate HasMatchers with HasMappingExt

Remove HasMappingExt and consolidate logic within HasMatchers. HasMatchers should handle both cases, when objects are ActiveFedora vs Valkyrie.

* 🧹 Fix Specs & add Valkyrie Specs

* 🧹 Fix Rubocop complaint

* 🧹 Address Valkyrie's determination of multiple?

* 🧹 Address permitted attributes

In Valkyrie, we use the schema to identify the permitted attributes. All
allowed attributes should be on the schema, so no additional attributes
should be required.

Also add a fallback for permitted attributes in case an ActiveFedora
model class goes through the ValkyrieObjectFactory. This supports the
case where we want to always force a Valkyrie resource to be created,
regardless of the model name given.

* 🧹 Update TODO comment

Adjust TODO message because referring to a handler that doesn't exist
anywhere is confusing. We may need to register steps for file sets once
the behavior is implemented.

---------

Co-authored-by: LaRita Robinson <laritakr@users.noreply.github.com>
Co-authored-by: Jeremy Friesen <jeremy.n.friesen@gmail.com>
Co-authored-by: LaRita Robinson <larita@scientist.com>