🎁 Config option for coercing a factory class name
Prior to this commit, once we had a factory class name we always called
`#constantize` on the string.

With this commit, we retain that default behavior by:

1. Introducing a function that does this thing
2. Exposing a configuration end-point for that function

Further, we add a looming Valkyrie convention; namely that for each
legacy named model (e.g. `Work`) there will likely be a corresponding
model with the Resrouce suffix (e.g. `WorkResource`).

Related to:

- https://github.com/scientist-softserv/hykuup_knapsack/issues/125