make find_by_source_identifier dynamic
Import a csv with child works. The forming of relationships is not working. Part of the problem is the find_by_bulkrax_identifier call.

From GBH, this used to be find_by_bulkrax_identifier which not all clients will configure as their source identifier. Instead we need to ask for the source identifier and use that for the sql query. This commit goes along with a PR from Hyku which currently has the find_by_source_identifier.rb files defined.

Issue:
- https://github.com/scientist-softserv/hykuup_knapsack/issues/128

Co-Authored-By: Kirk Wang <k3wang@gmail.com>