🧹 Make DeleteJob work wth new class method .find (#912)
* 🧹 Make DeleteJob work wth new class method .find

The DeleteJob previously was not working with the
old factory#find method because when it is doing a
delete action, the parsed_metadata does not get
generated like during a regular import.  Because
of this, the #search_by_identifier method fails to
find anything because we don't have a
`work_identifier` field which would have came from
the parsed_metadata.  So instead, we are using the
new class method .find which will take an id
(which we find on the raw_metadata) to find the
object.  We make sure to reindex and publish the
action to any relevant listeners.

* 🎁 Implement a #delete method for the ObjectFactory

This commit will add a delete method to the ObjectFactory and the
ValkyrieObjectFactory so we can avoid unnecessary conditionals.

* 🧹 Rework factories to implement delete method

This cuts down on the method chaining.