🐛 Fix exporter page for Bootstrap 3 applications
This commit will add the `hidden` class back in addition to the `d-none`
class.  This way older Hyku/Hyrax applications can still use it without
unintended effects.  Also, adding a turbolinks false to the edit
exporter link because the buttons on the form page was not working
without a refresh.